requirements.txt
django==5.1.1
mssql-django==1.6
pyodbc==5.2.0
sqlparse==0.5.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
celery==5.4.0
redis==5.0.8
reportlab==4.2.2
openpyxl==3.1.5
twilio==9.2.3
django-celery-beat==2.7.0
django-timezone-field==7.1
python-dateutil==2.9.0.post0